Added logic to load global PX as part of startup operations
