+++ Explorer/Assets/DCL/PluginSystem/DCL.Plugins.asmdef
@@ -130,7 +130,8 @@
         "GUID:543b8f091a5947a3880b7f2bca2358bd",
         "GUID:e5a23cdae0ef4d86aafc237a73280975",
         "GUID:ca4e81cdd6a34d1aa54c32ad41fc5b3b",
-        "GUID:1c5f9b69f95e40d18f11049c155eeaf8"
+        "GUID:1c5f9b69f95e40d18f11049c155eeaf8",
+        "GUID:8c4c611b27046bc4a848721d1fdd4a9f"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
