+++ Explorer/Assets/DCL/UserInAppInitializationFlow/DCL.UserInAppInitializationFlow.asmdef
@@ -30,7 +30,9 @@
         "GUID:8baf705856414dad9a73b3f382f1bc8b",
         "GUID:7175400a68914a45acecc9fb068de3b8",
         "GUID:ca4e81cdd6a34d1aa54c32ad41fc5b3b",
-        "GUID:3640f3c0b42946b0b8794a1ed8e06ca5"
+        "GUID:3640f3c0b42946b0b8794a1ed8e06ca5",
+        "GUID:f1eaef1b40a68e74cb90cbedebf57bbf",
+        "GUID:8c4c611b27046bc4a848721d1fdd4a9f"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
