+++ Explorer/Assets/DCL/UserInAppInitializationFlow/RealFlowLoadingStatus.cs
@@ -23,7 +23,8 @@ namespace DCL.UserInAppInitializationFlow
             ///     Player has teleported to the spawn point of the starting scene
             /// </summary>
             PlayerTeleported = 10,
-            Completed = 11,
+            LoadGlobalPXs = 11,
+            Completed = 12,
         }
 
         public static readonly Dictionary<Stage, float> PROGRESS = new (EnumUtils.GetEqualityComparer<Stage>())
@@ -39,6 +40,7 @@ namespace DCL.UserInAppInitializationFlow
             [Stage.OnboardingChecked] = 0.80f,
             [Stage.RealmRestarted] = 0.85f,
             [Stage.PlayerTeleported] = 0.95f,
+            [Stage.LoadGlobalPXs] = 0.99f,
             [Stage.Completed] = 1f,
         };
