+++ Explorer/Assets/Scripts/Global/Dynamic/DebugSettings.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1623b861e1c692c49900c7012328e140
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
