+++ Explorer/Assets/Scripts/Global/Dynamic/DebugSettings/Global.Dynamic.DebugSettings.asmdef
@@ -1,0 +1,16 @@
+{
+    "name": "Global.Dynamic.DebugSettings",
+    "rootNamespace": "",
+    "references": [
+        "GUID:702f733b4deb246808c6ce84d93b5c9c"
+    ],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
