+++ Explorer/Assets/Scripts/Global/Dynamic/DebugSettings/Global.Dynamic.DebugSettings.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 8c4c611b27046bc4a848721d1fdd4a9f
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
